tesseract/requirements.txt
python-dotenv
pytesseract
argparse
opencv-contrib-python
numpy
discord
requests
pillow
download
exif
